+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 1.1.3
+2022-01-09
+
+- Fix replace colors with chalk
+
+## 1.1.1
+2021-10-19
+
+- Fix in numeric minWidth calculation
+
 ## 1.1.0
 2021-10-19